+++ examples/middleware/src/env.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="astro/client" />
+declare namespace App {
+	interface Locals {
+		user: {
+			name: string;
+			surname: string;
+		};
+	}
+}