+++ packages/astro/src/core/render/index.ts
@@ -1,6 +1,11 @@
 export { createRenderContext } from './context.js';
 export type { RenderContext } from './context.js';
-export { getParamsAndProps, GetParamsAndPropsError, renderPage } from './core.js';
+export {
+	getParamsAndProps,
+	GetParamsAndPropsError,
+	getParamsAndPropsOrThrow,
+	renderPage,
+} from './core.js';
 export type { Environment } from './environment';
 export { createBasicEnvironment, createEnvironment } from './environment.js';
 export { loadRenderer } from './renderer.js';