+++ packages/astro/src/core/render/result.ts
@@ -50,6 +50,7 @@ export interface CreateResultArgs {
 	componentMetadata?: SSRResult['componentMetadata'];
 	request: Request;
 	status: number;
+	locals: App.Locals;
 }
 
 function getFunctionExpression(slot: any) {
@@ -131,7 +132,7 @@ class Slots {
 let renderMarkdown: any = null;
 
 export function createResult(args: CreateResultArgs): SSRResult {
-	const { markdown, params, pathname, renderers, request, resolve } = args;
+	const { markdown, params, pathname, renderers, request, resolve, locals } = args;
 
 	const url = new URL(request.url);
 	const headers = new Headers();
@@ -200,6 +201,7 @@ export function createResult(args: CreateResultArgs): SSRResult {
 				},
 				params,
 				props,
+				locals,
 				request,
 				url,
 				redirect: args.ssr