+++ packages/integrations/sitemap/test/fixtures/static/astro.config.mjs
@@ -1,0 +1,7 @@
+import { defineConfig } from 'astro/config';
+import sitemap from '@astrojs/sitemap';
+
+export default defineConfig({
+  integrations: [sitemap()],
+	site: 'http://example.com',
+})